test(docs): scope banner tests to config
